Moved registerUser() to RegisterActivity.kt ; Kdoc and refactoring
