+++ app/src/main/java/com/example/maraudersmap/UserController.kt
@@ -2,6 +2,12 @@ package com.example.maraudersmap
 
 import okhttp3.Response
 
+/**
+ * Provides methods to let the client communicate with the backend
+ * @author Julian Ertle
+ * @since 2022.11.30
+ */
+
 class UserController {
 
     /**
@@ -21,20 +27,9 @@ class UserController {
         userXTO.password = password
         userXTO.description = description
 
-
-            /*val userXml = """
-             <userXTO>
-                <username>${userXTO.username}</username>
-                <password>${userXTO.password}</password>
-                <description>${userXTO.description}</description>
-             </userXTO>
-            """.trimIndent()*/
-
-
         val server = ServerCommunicator()
         val response = server.postRequest("https://maraudersmap-ext.hhn.dev/api/v0.2/user", userXTO)
         return response
-
     }
 
     /**
@@ -49,17 +44,8 @@ class UserController {
         userXTO.username = username
         userXTO.password = password
 
-        /*val userXml = """
-             <userXTO>
-                <username>${userXTO.username}</username>
-                <password>${userXTO.password}</password> 
-             </userXTO>
-            """.trimIndent()*/
-
         val server = ServerCommunicator()
-
         val response = server.postRequest("https://maraudersmap-ext.hhn.dev/api/v0.2/user/login", userXTO)
-
         return response
     }
 }
